pipelines: update build pipeline to support verity boot.
Some boot optimizations have been made for SNP, where rootfs
is mounted using kernel's dm-verity feature. This brings the boot
time from ~15s to only a couple of seconds.

Additionally the makefile has been split into two, with the second
one depending only on `delta*.tar.gz` and a `tar2ext4` tool. As a
result, the VMGS template will be updated to utilize the second
`Makefile.bootfiles`, meaning that the `delta*.tar.gz` is expected
to have the final delta and we'll not be re-assembling it in the
pipeline. As a result, we need to make sure that kernel init script
is in its final form and has the initial policy stance set to "deny".
The `startup.sh` needs to be hacked to include those arguments to
`gcs` binary as done in this PR.

Signed-off-by: Maksim An <maksiman@microsoft.com>