scsi: Take optional guest path for mount
Currently the SCSI mount manager will generate a new path for each new
mount, based on a format string that it is instantiated with. However,
it turns out some code in the GCS (e.g. sandbox mounts) assumes that the
container scratch is mounted at a certain path. The long-term best
solution here is probably to pass what paths to use explicitly to the
GCS, but that would be more impactful. We need a more contained fix.

This commit addresses the issue by allowing an optional guest path to be
given for a SCSI mount. The mount manager has been changed as follows:
- If a guest path is not supplied: The mount can re-use (refcount++) any
  existing mount with the same controller/lun/options. If a new mount is
  created, the mount manager will generate a path for it.
- If a guest path is supplied: The mount can re-use (refcount++) any
  existing mount with the same controller/lun/guestpath/options. If a
  new mount is created, the mount manager will use the supplied path for
  it.

Accordingly, code calling into the mount manager has been updated to
pass an empty string for the guest path. The exception to this is the
LCOW layer mounting code, which will pass an explicit guest path for the
scratch disk. As far as I know, WCOW does not depend on a specific path
for its scratch disk.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>
(cherry picked from commit c1d649299c57828e494dc743c99e32a53d22a0b1)