Require various regex.match in framework.rego to be full match
It seems to me that for 9p mounts from the host into a UVM, we are only supposed
to mount to exactly

    ^/run/gcs/c/<containerID>/mounts/m[0-9]+$

i.e. /run/gcs/c/.../m? is not just a prefix check.  The only place which uses
this prefix is allocateLinuxResources, and it doesn't try to mount to anything
under the m<number> directory.  Hence we should make the regex match there be
full string match too.

Combined with the fact that 9p mountpoints are already checked for duplicate,
this prevents any “mounting on top of symlinks” tricks from the host.

Another example, for mount source, the policy usually has:

    "mounts": [
      {
        "destination": "/etc/resolv.conf",
        "options": [
          "rbind",
          "rshared",
          "rw"
        ],
        "source": "sandbox:///tmp/atlas/resolvconf/.+",
        "type": "bind"
      }
    ],

and is intended to enforce that when starting the container, the source for the
/etc/resolv.conf mount must come from /tmp/atlas/resolvconf/ within the
sandboxMounts. Similar policies will be generated for file mounts:

      {
        "destination": "/mnt/volume",
        "options": [
          "rbind",
          "rshared",
          "rw"
        ],
        "source": "sandbox:///tmp/atlas/azureFileVolume/.+",
        "type": "bind"
      },

This commit changes these cases to so that we use the anchored pattern,
effectively enforcing a full match.

Fixes: https://msazure.visualstudio.com/One/_workitems/edit/33064760
Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>