Enforce container/sandbox IDs can't contain any "dangerous" characters.
Since these IDs are used to construct various paths (mount dir, path for
resolv.conf, scratch path (containerScratchPathInUVM in lcow.go)), there is
potential for path traversal attack here.  We check that it can't be something
weird for confidential containers (while still allowing anything the host passes
if we're not in confidential mode, to not accidentally break other
dependencies).

Aside from CreateContainer, we also check for this in modify*Settings in case
the host crafts a request with a malformed ID later on.

Since the functional tests uses names like
TestContainerExecLCOW-1df570f2-container, we can't enforce that this must
strictly be a hex or a UUID.

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>