Enforce mount policy for read-write disk mounts
The main purpose of this is to prevent mounting host-controlled, non-encrypted
filesystems.  Combined with the ability to mount to anywhere, this results in
code execution from the host into the guest.  Unencrypted and
non-integrity-checked disks on its own also runs the risk of kernel filesystem
bugs.

While this commit does not yet prevent mounting to arbitrary paths, it makes
exploiting this much more difficult now as all the host can do is mount an empty
directory on top of things.

Using the reproducer in the bug report, we get the following deny message:
  {"decision":"deny","input":{"encrypted":false,"ensureFilesystem":false,"filesystem":"","readonly":false,"rule":"mount_device","target":"/bin/"},"reason":{"errors":["ensureFilesystem must be set on rw device mounts","rw device mounts uses a filesystem that is not allowed","unencrypted scratch not allowed, device to be mounted must be encrypted"]}}

Fixes: https://msazure.visualstudio.com/One/_workitems/edit/33144273
Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>