Enforce valid rootfs, bundle and scratch paths in CreateContainer
If the host can control the rootfs path, by using otherwise legitimate 9p
mounts, this lets it take over the container. Similar exploits might be possible
by controlling ScratchDirPath or OCIBundlePath in this request as well, so we
enforce that those paths are as expected too.

Since in a previous commit we already ensure that 9pfs can only be mounted to
/run/gcs/c/.../mounts, this means that it is no longer possible to use 9p mounts
to exploit this, and nor should it be possible to use e.g. something in
sandbox://.

A later commit should also ensure that disks cannot be mounted to arbitrary
paths too.

Error message example:
  time="2025-06-09T15:34:38Z" level=fatal msg="starting the container \"fdb12ddbbfdeb1ee990abe892a06ac80304e5298c48d7eae2845c201e72efd5b\": rpc error: code = Unknown desc = failed to create containerd task: failed to create shim task: failed to create container fdb12ddbbfdeb1ee990abe892a06ac80304e5298c48d7eae2845c201e72efd5b: guest RPC failure: OCISpecification.Root.Path \"/run/gcs/c/f3c2e64041edd6aa1c7f20c15c2bed0d2afffabb3b22aceac391d8f43d9fc567/mounts/m0\" must equal expected \"/run/gcs/c/fdb12ddbbfdeb1ee990abe892a06ac80304e5298c48d7eae2845c201e72efd5b/rootfs\": unknown"

Fixes: https://msazure.visualstudio.com/One/_workitems/edit/33205622
Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>