test/functional: Fix TestGetProperties_WithPolicy
This test is currently broken already, due to VPMem multimapping not being
disabled, and also due to missing required environment variables.  In addition
to fixing that, this commit also make it work with the latest changes, by using
proper containerID, and disabling VPMem altogether (as it will fail the mount
target check).