Do not allow overriding filesystem type for read-only mounts on confidential
At least on confidential mode, read only disks are only supposed to be ext4, and
in fact if it isn't, we will fail to read the verity info, and error out
earlier.

However, currently the host can specify a filesystem even when we're mounting a
dm-verify protected volume.  This presents a risk for exploits (but is not
currently exploitable), for example the host could specify virtiofs, and prepare
the correct vhost socket with tags like /dev/mapper/dm-verify-..., and be able
to override a container layer.

This is not currently exploitable thanks to the fact that we try to mount with
the noload option first, and mounting would not continue if that fails.  This
option will prevent mounting virtiofs, 9p, overlay, etc.

Note that the host can also specify its own options, but those options are only
applied if we get through the first "mount with noload" stage.

This commit prevents the host from specifying anything other than ext4, thus
eliminating this risk.