Use a separate enforcement point for read-write device mounts
Mahati suggested below to add a new enforcement point rw_mount_device, instead
of adding an input.readonly to the existing mount_device.  This commit does
that, and also bump the API version and set introducedVersion for the new rule,
so that old policies will fallback to the "default" for this enforcement point,
which in this commit is defined to allow.

This also has the benefit that handwritten policies that does slightly different
things in mount_device would not break.  Since previously written policies would
not have handled the scratch mount in mount_device, using a separate rule for
read-write mounts is less likely to break those.

However, doing it this way means that existing policies does not get this rw
mountpoint protection.  To remedy that, the next commit will add a
"use_framework" mechanism to pass through new enforcement points to the
framework by default, and use it for rw_mount_device instead of defaulting to
allow.

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>