guest/scsi: Cleanup mountpoint/dm-crypt device on scratch disk dm setup / mount error
Previously, if the dm-crypt setup fails, the mountpoint is not deleted due to
(accidental?) shaowing of the `err` variable.  Also, on mount failures, the
dm-crypt device is not cleaned up.  This leaks the dm-crypt device (and prevent
another mount of a scratch disk with the same scsi controller and lun from
succeeding), and also leaves the device in an in-use state which is undesirable
given that the host will immediately try to detach it.

While those failures should be rare (and in practice, since confidential
containers use shared scratch, azcri will normally never try another one if the
scratch mount for the pause container fails, as that would be a UVM startup
failure), the repro case for 6b5b89346 ("hcsv2/uvm: Rollback policy state on
mount errors") has showed that leaving things lying around can make exploits
easier (e.g. in that particular case, if gcs did not clean up the mountpoint, we
would have been able to mount the overlay and start a container with missing
layers).

This also makes the policy rollback we're about to do on mount failures more
correct.

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>