hcsv2/uvm: Deny removal of in-use overlays for confidential containers
While any mount changes from gcs, which lives in the init namespace, wouldn't
actually affect any container that has already started anyway, due to the
container being in a separate mount namespace, the host trying to unmount an
overlay that is in use is still incorrect, and we harden ourselves by preventing
that.

Together with hostMounts, this also prevents us trying to close down a dm device
that is in use by a scratch / layer (even though Linux would EBUSY that anyway).
If we allow unmounting the in-use overlay, then the host can do that first, then
go on to ask us to unmount the scratch / layer and close the dm.

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>