hcsv2/uvm: Only use hostMounts for confidential containers.
[cherry-picked from d1dbec46c86d08d9babf5fcd1b0d8445e7d878e4

note: since SetConfidentialUVMOptions is now refactored into
securityOptions.SetConfidentialOptions, that function now no longer have access
to host. Thus the original change to it which initializes newHostMounts is now
moved directly to modifyHostSettings]

Since we're placing additional restrictions on when unmounts are allowed, we
ensure that the impact of this change is scoped to confidential containers only.

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>