Merged PR 12916199: hcsv2/uvm: Check that OCISpecification.Hooks is nil for confidential
[cherry-picked from 73a7151ad8f15aad9f0e4eef6b86ed1a1a5a0953]

This fixes a vulnerability where the host can provide arbitrary commands as
hooks in the OCISpecification, which will get executed when passed to runc.

Normally, ociSpec.Hooks is cleared explicitly by the host, and so we should
never get passed hooks from a genuine host.  The only case where OCI hooks would
be set is if we are running a GPU container, in which case guest-side code in
GCS will add hooks to set up the device - see addNvidiaDeviceHook.

This restriction will also apply to the hook set this way by ourselves, however
this is fine for now as we do not support Confidential GPUs.  When we do support
GPU confidential containers, we will need to pull this enforcement into Rego,
and allow the policy to determine whether the hook should be allowed (along with
enforcing which devices are allowed to be exposed to the container).

Even after this fix, we still have unenforced fields in the OCISpecification
that we receive from the host, that may allow the host to, for example, add
device nodes, and we should harden this further in a future release.

Tested using the following host-side reproduction patch:

```diff
diff --git a/internal/hcsoci/hcsdoc_lcow.go b/internal/hcsoci/hcsdoc_lcow.go
index db94d73df..a270c13ac 100644
--- a/internal/hcsoci/hcsdoc_lcow.go
+++ b/internal/hcsoci/hcsdoc_lcow.go
@@ -94,6 +94,22 @@ func createLinuxContainerDocument(ctx context.Context, coi *createOptionsInterna
 		return nil, err
 	}

+	isSandbox := spec.Annotations[annotations.KubernetesContainerType] == "sandbox"
+
+	if !isSandbox {
+		if spec.Hooks == nil {
+			spec.Hooks = &specs.Hooks{}
+		}
+		timeout := 5000000
+		spec.Hooks.StartContainer = append(spec.Hooks.StartContainer, specs.Hook{
+			Path:    "/bin/sh",
+			Args:    []string{"/bin/sh", "-c", "echo hacked by hooks > /hacked.txt"},
+			Env:     []string{"PATH=/usr/local/bin:/usr/bin:/bin:/sbin"},
+			Timeout: &timeout,
+		})
+		log.G(ctx).Info("Injected hook into OCISpec")
+	}
+
 	log.G(ctx).WithField("guestRoot", guestRoot).Debug("hcsshim::createLinuxContainerDoc")
 	return &linuxHostedSystem{
 		SchemaVersion:    schemaversion.SchemaV21(),
```

Non-confidential scenarios should not be affected.

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>