rego: Block loading of fragments with unsafe namespaces
For an arbitrary fragment offered by the host, we need to interpret the fragment
code as a new Rego module in order for the framework to check it.  However,
currently we do not check that the namespace of the fragment, which is used as
the namespace of the module, does not conflict with any of our own namespaces
(e.g. framework, policy, or api).  This means that as soon as we load it for
policy checking, a malicious fragment could override enforcement points defined
in the framework, including "load_fragment", and allow itself in, even if we
would otherwise have denied it because of wrong issuer, etc.

This can be tested with the below fragment:

	package framework

	svn := "1"
	framework_version := "0.3.1"

	load_fragment := {"allowed": true, "add_module": true}
	enforcement_point_info := {
		"available": true,
		"unknown": false,
		"invalid": false,
		"version_missing": false,
		"default_results": {"allowed": true},
		"use_framework": true
	}

	mount_device := {"allowed": true}
	rw_mount_device := {"allowed": true}
	mount_overlay := {"allowed": true}
	create_container := {"allowed": true, "env_list": null, "allow_stdio_access": true}
	unmount_device := {"allowed": true}
	rw_unmount_device := {"allowed": true}
	unmount_overlay := {"allowed": true}
	exec_in_container := {"allowed": true, "env_list": null}
	exec_external := {"allowed": true, "env_list": null, "allow_stdio_access": true}
	shutdown_container := {"allowed": true}
	signal_container_process := {"allowed": true}
	plan9_mount := {"allowed": true}
	plan9_unmount := {"allowed": true}
	get_properties := {"allowed": true}
	dump_stacks := {"allowed": true}
	runtime_logging := {"allowed": true}
	scratch_mount := {"allowed": true}
	scratch_unmount := {"allowed": true}

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>