rego: Do not load fragment as Rego code if issuer or feed is not trusted.
This further reduces the attack surface by not loading untrusted Rego, which
might attempt to exploit other vulnerabilities in the Rego parser / Open Policy
Agent.

To do this, load_fragment is now split into two modes - first called with
fragment_loaded == false, in which case it just checks the issuer and feed, then
if that passes it is called again with fragment_loaded == true, and can then
access the fragment namespace to check the SVN and other fields.

Doing it this way, any existing policies which either forwards load_fragment to
the framework, or allows it unconditionally, would stay behaving the same.

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>