regopolicyinterpreter: Add Set metadata (in additional to string maps)
In implementing the fragment parameters feature, we need a way to store, for
each issuer and feed tuple, which parameters should be used for fragments
matching it.  Consider this list of fragments, which a parent fragment might
define:

fragments := [
    {
        "feed": "mcr.microsoft.com/maa/enclavehost",
        "issuer": "did:x509:0:sha256:...",
        "includes": [
            "containers",
            "fragments"
        ],
        "minimum_svn": "1",
        "parameters": {
            "region": "australiacentral2",
            "cloud": "Public"
        }
    },
    // ...
]

When we load the fragment containing this, we need to iterate through its
data.<namespace>.fragments array, and for each entry, append the parameters
object to an array that is keyed by the issuer and feed, since we can have
multiple such fragment import entries for the same issuer and feed, but with
different parameters.

This basically means that we need the equivalent of the following code in Rego,
which I've failed to come up with a way to write:

    for f in fragment_fragments {
        issuer = data.metadata.issuers[f.issuer] or {}
        feed = issuer.feeds[f.feed] or {}
        feed.parameters = feed.parameters or []
        feed.parameters = array_append(feed.parameters, f.parameters)
        issuer.feeds[f.feed] = feed
        data.metadata.issuers[f.issuer] = issuer
    }

While we can dynamically lookup or store based on a key that is from a variable
access, and we can send multiple metadata updates, those updates cannot express
the semantic of "merging" objects or arrays.

To make this simpler, we introduce a new metadata data type - "set", which
supports storing an unordered list of arbitrary objects (which themselves might
be either a string, or an object containing key-value pairs), that can be
inserted to via metadata operations one at a time.  This means that we can
simply append to the metadata update operations array once per each fragment
import statement, and the outcome would be the union of all the parameter
objects.  This is also very easy to query in Rego, given an issuer and feed:

    parameters := [
        fp.parameters |
            fp = data.metadata.fragment_parameters[_]
            fp.issuer == input.issuer
            fp.feed == input.feed
    ]

It is likely that the existing code that extracts included containers/fragments
from a fragment can be simplified by using this feature, but that is outside of
the scope of this PR.

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>