rego: Implement fragment parameters
This commit implements the support for the "parameters" feature for fragments.
This is achieved by having the framework extract the parameters object from
fragment import statements, and storing them in a set, tagged with the
applicable (issuer, feed) pair.  On future fragment loads, if the fragment's
issuer and feed pair matches with any previous entry from this set, the
parameters will be passed to the fragment via an automatically injected object
added to the end of the fragment's Rego source (__fragment_parameters).

(Note that in reality, we need to combine this set with the import statements
defined in the main policy. c.f. candidate_fragments.  This is done in
fragment_parameters_for)

In order for fragments to use this parameter object, it is expected that all
fragments will now have an additional "stub" inserted at runtime to define the
parameter() function, which will, under the hood, reference a "hidden" variable
__fragment_parameters, also inserted dynamically at runtime, containing the
actual parameter values.

If multiple fragment parameters are specified, all combinations of values are
considered "allowed", and therefore the fragment injection is repeated, passing
in different parameter combinations, one for each combination.  This means that
if a fragment defines, for example:

	containers := [
		{
			...
			"env_rules": [
				{
					"name": "SOME_KEY",
					"name_strategy": "string",
					"value": parameter("my_param"),
					"value_strategy": "string"
				}
			]
		}
	]

and we have two fragment import statement for this fragment, e.g.:

fragments := [
	{
		"feed": ...,
		"issuer": ...,
		"minimum_svn": ...,
		"parameters": {
			"my_param": "value1"
		}
	},
	{
		"feed": ...,
		"issuer": ...,
		"minimum_svn": ...,
		"parameters": {
			"my_param": "value2"
		}
	}
]

Then the container can be started with either SOME_KEY=value1 or
SOME_KEY=value2.

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>
---

Changes:
- Fix missing [_]
- fix wrong usage of defer
- Inject fragment parameter function definitions at runtime:

  We must inject this at runtime so as to avoid having to support this exact
  implementation (eg __fragment_parameters[name]) of fetching the parameters
  forever (as it will essentially be hard-coded in the generated policy).

- Move the actual implementation of parameter() into the framework so that
  fragments doesn't have to say 'import future.keywords.every', 'import
  future.keywords.in'.

- Use indirection with default when accessing fragment.parameters to not break
  older fragments.

  If we don't do this, loading a fragment without the parameters definition
  would fail due to "unsafe" rego variable references.

- Fix broken tests caused by the new extract_parameter framework function