guest: Restrict plan9 share names to digits only on Confidential mode
Currently the host can pass in a share name with injected option in it, e.g.
"123,cache=loose".  While this is currently probably harmless, it's still an
risk and so we should block these kind of mount option injections.

In Linux, this is parsed by v9fs_parse_options. It basically scans until the
next ',', and it doesn't matter whether we add quotes.  In hcsshim, all plan9
mounts go through AddPlan9 on the host side, and that function uses a number as
the share name.  Therefore, we will simply restrict the share name to be digits
only.

Test:
	+++ b/internal/uvm/plan9.go
	@@ -90,7 +90,7 @@ func (uvm *UtilityVM) AddPlan9(ctx context.Context, hostPath string, uvmPath str
				RequestType:  guestrequest.RequestTypeAdd,
				Settings: guestresource.LCOWMappedDirectory{
					MountPath: uvmPath,
	-				ShareName: name,
	+				ShareName: name + ",cache=loose",
					Port:      plan9Port,
					ReadOnly:  readOnly,
				},

Output:
	failed to share directory C:\lcow_info\ into UVM: rpc error: code = Unknown desc = guest modify: guest RPC failure: invalid plan9 share name "1,cache=loose": must match regex "^[0-9]+$"

Closes: https://msazure.visualstudio.com/One/_workitems/edit/34370380

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>