hcsv2/uvm: Restrict SCSI mount options in confidential mode
In C-LCOW, we do not want to host to be able to arbitrarily control mount
options.  Currently there are two possible ways mount options might be
specified by the host:

1. For read-only mounts (image layers), option "ro" is specified (see
   addLCOWLayer).
2. If the OCI spec passed by containerd contains physical/virtual disk mounts,
   it might contain mount options, and hcsshim would pass this through to GCS (see
   allocateLinuxResources).

We can allow 1 (and in fact, require it to be consistent with the readOnly field
in the request), and today C-LCOW does not support external disk mounts, and so
we can reject any other mount options passed via route 2.

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>