hcsv2/uvm, rego: Enforce OCI spec does not contain any LinuxDevices
Modify the create_container enforcement point to take in an extra "devices"
input for Rego to enforce on.

Even though Rego gets the devices array as input, for now we will just not allow
any devices in framework.rego, and there is no way for a normal policy to allow
them, but this may be changed in the future, where the container definition in
the policy may contains a "devices" array defaulting to [].

We should let the policy enforcer ignore any dynamically added devices to
privileged containers, since if the policy allows privileged, adding devices to
/dev should be expected.  We should also allow /dev/sev-guest unconditionally,
without the policy needing to have a rule for it.  In order to implement this,
we simply take a copy of the devices array before we populate it with GCS-found
devices and sev-guest, and pass that copy to the enforcer.

Testing:
- Non-privileged and privileged containers work as expected, with the correct
  devices.
- Modifying the host to inject malicious devices to workload container,
  privileged containers gets policy denial ("devices not supported"), whereas
  non-privileged containers gets "guest RPC failure: no such file or directory:
  unknown" due to GCS trying to resolve the device's major/minor number for
  cgroup.
- If the malicious device is injected to the sandbox container instead, we get
  policy denials.

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>