regopolicyinterpreter: Add copyRegoMetadata
We will later need to copy the metadata in order to implement revertable
sections.

This extra function is necessary due to the fact that
map[string]map[string]interface{} is not convertable to map[string]interface{}
(or back), so the existing copy function cannot be used.

An alternative is to turn the deep copy functions into one generic function, but
reviewer was against it.

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>