hcsv2: Change hostMounts to be caller-locked.
Since the expected usage of this struct expects caller to undo operations in
case of failure, it makes more sense to expect that the caller hold the lock
throughout, until it has either committed the operation or undone it.  This
prevents accidental misuse, although in practice this struct is unlikely to be
called from different threads anyway due to sequential bridge message processing
in confidential containers.

Locking hostMounts this way effectively means that mount/unmount operations are
always single-threaded.  However, this is the case in confidential containers
anyway due to the sequential message processing, and on non-confidential
containers hostMounts isn't used for now.

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>