hcsv2/uvm: Revert metadata and block all further mount/unmount/CreateContainer if any unmount fails
Unlike mounts, if a unmount fails (either because the unmount itself fails, or
in the case of scratch disks, the unmount succeed but dm-crypt fails to close
the device), we cannot safely revert the Rego metadata, as we will not be able
to properly "undo" the unmount if it has successfully happened.  Even if the
unmount itself fails, this is still a highly unexpected state, and we might not
want to continue using the device as usual.

One possible solution here is just to not revert the Rego metadata, and let rego
pretend that the unmount succeeded.  Doing so for SCSI disks and overlay mounts
is probably OK, but not for 9p as we might inadvertently be allowing the host to
trick us into mounting on top of an existing host-controlled 9p system (which
might then lead to possible symlink-related exploits?).

However, since we really do not expect unmounts to fail (aside from policy
enforcement failure, or if the host passes an invalid mount point), if it does
then we protect ourselves by bailing out of all further mounts and unmounts
(including overlay).  We prevent ourselves from starting new containers as well
to protect against using a "broken" overlay.

Note that if the policy denies such a unmount, we won't end up in this broken
state, and will just return an error as usual.

Fixes: https://msazure.visualstudio.com/One/_workitems/edit/33232631
Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>