rego: Implement platform_rules support
This currently support containers[_].env_rules and containers[_].mounts.  If
multiple platform_rules are defined, a container matching either one can be
started (but in a consistent manner - e.g. if two platforms have different
environment variables or mounts, a container can't "mix and match" between
them).

In order to achieve the above consistency, we "patch" the container objects
instead of adding logic to e.g. env_rule_ok or envList_ok.  This also means that
the error_objects of a denial message will reflect the platform rules inserted.

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>