rego: Fix create_container enforcement deny message missing input.rule field
Currently we only add input.rule to the original input, not the redacted input.
This results in the case of create_container not having the "rule" field in the
final deny message, but other enforcement points do have it since in those cases
the redactSensitiveData return the original input map.

Signed-off-by: Tingmao Wang <tingmaowang@microsoft.com>